+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes
 - Let MagikVisitorContext handle Scope building
 - Add ${tag} to MessageFormatReporter
 - Changes to MagikGrammar
+- Give FormattingCheck proper key
 
 
 0.1.4 (2019-07-26)