+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes
 - Changes to MagikGrammar
 - Give FormattingCheck proper key
 - Fix magik-lint --show-checks
+- Fix memleak in UnusedVariableCheck
 
 
 0.1.4 (2019-07-26)